'Add GlobalDenseBaseModule'
