+++ models/core_modules.py
@@ -25,6 +25,165 @@ def MLP(channels, activation=ReLU()):
         for i in range(1, len(channels))
     ])
 
+################## BASE MLP BUILT USING CONV_2D #####################
+
+
+class SharedMLP(nn.Sequential):
+    def __init__(
+        self,
+        layers,
+        bn=False,
+        activation=nn.ReLU(inplace=True),
+        preact=False,
+        first=False,
+        name="",
+    ):
+        # type: (SharedMLP, List[int], bool, Any, bool, bool, AnyStr) -> None
+        super(SharedMLP, self).__init__()
+
+        for i in range(len(layers) - 1):
+            self.add_module(
+                name + "layer{}".format(i),
+                Conv2d(
+                    layers[i],
+                    layers[i + 1],
+                    bn=(not first or not preact or (i != 0)) and bn,
+                    activation=activation
+                    if (not first or not preact or (i != 0))
+                    else None,
+                    preact=preact,
+                ),
+            )
+
+
+class _ConvBase(nn.Sequential):
+    def __init__(
+        self,
+        in_size,
+        out_size,
+        kernel_size,
+        stride,
+        padding,
+        dilation,
+        activation,
+        bn,
+        init,
+        conv=None,
+        norm_layer=None,
+        bias=True,
+        preact=False,
+        name="",
+    ):
+        super(_ConvBase, self).__init__()
+
+        bias = bias and (not bn)
+        conv_unit = conv(
+            in_size,
+            out_size,
+            kernel_size=kernel_size,
+            stride=stride,
+            padding=padding,
+            dilation=dilation,
+            bias=bias,
+        )
+        init(conv_unit.weight)
+        if bias:
+            nn.init.constant_(conv_unit.bias, 0)
+
+        if bn:
+            if not preact:
+                bn_unit = norm_layer(out_size)
+            else:
+                bn_unit = norm_layer(in_size)
+
+        if preact:
+            if bn:
+                self.add_module(name + "normlayer", bn_unit)
+
+            if activation is not None:
+                self.add_module(name + "activation", activation)
+
+        self.add_module(name + "conv", conv_unit)
+
+        if not preact:
+            if bn:
+                self.add_module(name + "normlayer", bn_unit)
+
+            if activation is not None:
+                self.add_module(name + "activation", activation)
+
+
+class Conv1d(_ConvBase):
+    def __init__(
+        self,
+        in_size,
+        out_size,
+        kernel_size=1,
+        stride=1,
+        padding=0,
+        dilation=1,
+        activation=nn.ReLU(inplace=True),
+        bn=False,
+        init=nn.init.kaiming_normal_,
+        bias=True,
+        preact=False,
+        name="",
+        norm_layer=nn.BatchNorm1d,
+    ):
+        # type: (Conv1d, int, int, int, int, int, int, Any, bool, Any, bool, bool, AnyStr, _BNBase) -> None
+        super(Conv1d, self).__init__(
+            in_size,
+            out_size,
+            kernel_size,
+            stride,
+            padding,
+            dilation,
+            activation,
+            bn,
+            init,
+            conv=nn.Conv1d,
+            norm_layer=norm_layer,
+            bias=bias,
+            preact=preact,
+            name=name,
+        )
+
+
+class Conv2d(_ConvBase):
+    def __init__(
+        self,
+        in_size,
+        out_size,
+        kernel_size=(1, 1),
+        stride=(1, 1),
+        padding=(0, 0),
+        dilation=(1, 1),
+        activation=nn.ReLU(inplace=True),
+        bn=False,
+        init=nn.init.kaiming_normal_,
+        bias=True,
+        preact=False,
+        name="",
+        norm_layer=nn.BatchNorm2d,
+    ):
+        # type: (Conv2d, int, int, Tuple[int, int], Tuple[int, int], Tuple[int, int], Tuple[int, int], Any, bool, Any, bool, bool, AnyStr, _BNBase) -> None
+        super(Conv2d, self).__init__(
+            in_size,
+            out_size,
+            kernel_size,
+            stride,
+            padding,
+            dilation,
+            activation,
+            bn,
+            init,
+            conv=nn.Conv2d,
+            norm_layer=norm_layer,
+            bias=bias,
+            preact=preact,
+            name=name,
+        )
+
 ################## BASE CONVOLUTION #####################
 
 
@@ -214,9 +373,7 @@ class GlobalBaseModule(torch.nn.Module):
     def forward(self, data):
         batch_obj = Batch()
         x, pos, batch = data.x, data.pos, data.batch
-        pos_flipped = pos.transpose(1, 2).contiguous()
-
-        x = self.nn(torch.cat([x, pos_flipped], dim=1))
+        x = self.nn(torch.cat([x, pos], dim=1))
         x = self.pool(x, batch)
         batch_obj.x = x
         batch_obj.pos = pos.new_zeros((x.size(0), 3))
@@ -228,18 +385,17 @@ class GlobalBaseModule(torch.nn.Module):
 class GlobalDenseBaseModule(torch.nn.Module):
     def __init__(self, nn, aggr='max'):
         super(GlobalDenseBaseModule, self).__init__()
-        self.nn = MLP(nn)
+        self.nn = SharedMLP(nn)
 
     def forward(self, data):
         batch_obj = Batch()
         x, pos, batch = data.x, data.pos, data.batch
-        import pdb
-        pdb.set_trace()
-        x = self.nn(torch.cat([x, pos], dim=1))
-        x = self.pool(x, batch)
+        pos_flipped = pos.transpose(1, 2).contiguous()
+        x = self.nn(torch.cat([x, pos_flipped], dim=1).unsqueeze(-1))
+        x = x.squeeze().max(-1)[0]
         batch_obj.x = x
         batch_obj.pos = pos.new_zeros((x.size(0), 3))
-        batch_obj.batch = torch.arange(x.size(0), device=batch.device)
+        batch_obj.batch = torch.arange(x.size(0), device=x.device)
         copy_from_to(data, batch_obj)
         return batch_obj
